Break classes into separate modules, add comments
