+++ src/Augraphy/Augmentations/dirtyrollers.py
@@ -1,0 +1,146 @@
+################################################################################
+# File: dirtyrollers.py
+#
+# Class: DirtyRollersAugmentation
+#
+# Description: This file contains a class defining an Augmentation which
+#              emulates an effect created by certain document scanners.
+################################################################################
+
+
+
+################################################################################
+# Imports
+################################################################################
+
+import cv2
+import numpy as np
+import random
+
+
+
+################################################################################
+# Class Definition
+################################################################################
+
+class DirtyRollersAugmentation(Augmentation):
+  def __init__(self, line_width_range=(8, 12), probability=0.5):
+    super().__init__(probability=probability)
+    self.line_width_range = line_width_range
+
+  # Constructs a string representation of this Augmentation.
+  def __repr__(self):
+    return f"DirtyRollersAugmentation(line_width_range={self.line_width_range}, probability={self.probability})"
+
+  def apply_scanline_mask(self, img, mask, meta_mask):
+    if (random.choice([True, False])):
+      return self.apply_scanline_mask_v2(img, mask, meta_mask)
+    else:
+      return self.apply_scanline_mask_v1(img, mask, meta_mask)
+
+  def apply_scanline_mask_v2(self, img, mask, meta_mask):
+    mask = self.transform(self.apply_scanline_metamask_v2, mask, meta_mask)
+    update_lambda = lambda x, y: min(255, x + (x * (1-(y/100))))
+    update = np.vectorize(update_lambda)
+    return update(img, mask)
+
+  def apply_scanline_metamask_v2(self, img, mask):
+    update_lambda = lambda x, y: max(0, x - (x * (1-(y/100))))
+    update = np.vectorize(update_lambda)
+    return update(img, mask)
+
+  def apply_scanline_mask_v1(self, img, mask, meta_mask):
+    mask = self.transform(self.apply_scanline_metamask_v1, mask, meta_mask)
+    update_lambda = lambda x, y: max(0, x - (x * (1-(y/100))))
+    update = np.vectorize(update_lambda)
+    return update(img, mask)
+
+  def apply_scanline_metamask_v1(self, img, mask):
+    update_lambda = lambda x, y: min(99, x + (x * (1-(y/100))))
+    update = np.vectorize(update_lambda)
+    return update(img, mask)
+
+  # Applies the Augmentation to input data.
+  def __call__(self, data, force=False):
+    if (force or self.should_run()):
+      image = data['post'][-1].result
+      line_width = random.randint(self.line_width_range[0], self.line_width_range[1])
+      rotate = random.choice([True, False])
+
+      if (rotate):
+        image = cv2.rotate(image, cv2.ROTATE_90_CLOCKWISE)
+
+      image = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
+      mask = self.transform(self.create_scanline_mask, image.shape[1], image.shape[0], line_width)
+      meta_mask = self.transform(self.create_scanline_mask, image.shape[1], image.shape[0], line_width * random.randint(10, 25))
+      image = self.transform(self.apply_scanline_mask, image, mask, meta_mask).astype("uint8")
+      image = cv2.cvtColor(image, cv2.COLOR_GRAY2BGR)
+
+      if (rotate):
+        image = cv2.rotate(image, cv2.ROTATE_90_COUNTERCLOCKWISE)
+
+      metadata = dict()
+      metadata['mask'] = mask
+      metadata['metamask'] = meta_mask
+      data['post'].append(self, image, metadata)
+
+  def create_scanline_mask(self, width, height, line_width):
+    grad_list = list()
+
+    # Create Standard Bar
+    grad_high_pct = random.randint(86, 99)
+    grad_low_pct = random.randint(70,85)
+
+    grad_dec = (grad_low_pct-grad_high_pct)/(line_width)
+    grad_grid = np.mgrid[grad_high_pct:grad_low_pct:grad_dec]
+    grad_grid = np.hstack((grad_grid, np.flip(grad_grid)))
+    grad_grid = np.tile(grad_grid,(height,1))
+    grad_list.append(grad_grid)
+
+    # Create Standard Bar with Wide Dark
+    grad_dec = (grad_low_pct-grad_high_pct)/(line_width)
+    grad_grid = np.mgrid[grad_high_pct:grad_low_pct:grad_dec]
+    grad_center = np.full((random.randint(1, 6)), grad_low_pct)
+    grad_grid = np.hstack((grad_grid, grad_center, np.flip(grad_grid)))
+    grad_grid = np.tile(grad_grid,(height,1))
+    grad_list.append(grad_grid)
+
+    # Create Standard Bar with Wide Light
+    grad_dec = (grad_low_pct-grad_high_pct)/(line_width)
+    grad_grid = np.mgrid[grad_high_pct:grad_low_pct:grad_dec]
+    grad_exterior = np.full((random.randint(1, 6)), grad_high_pct)
+    grad_grid = np.hstack((grad_grid, np.flip(grad_grid), grad_exterior))
+    grad_grid = np.tile(grad_grid,(height,1))
+    grad_list.append(grad_grid)
+
+    # Create Standard Bar with Lower Dark
+    grad_high_pct += min(100, random.randint(-3, 3))
+    grad_low_pct -= random.randint(5, 8)
+    grad_dec = (grad_low_pct-grad_high_pct)/(line_width)
+    grad_grid = np.mgrid[grad_high_pct:grad_low_pct:grad_dec]
+    grad_grid = np.hstack((grad_grid, np.flip(grad_grid)))
+    grad_grid = np.tile(grad_grid,(height,1))
+    grad_list.append(grad_grid)
+
+    # Create Standard Bar with Low Dark and Wide Dark
+    grad_dec = (grad_low_pct-grad_high_pct)/(line_width)
+    grad_grid = np.mgrid[grad_high_pct:grad_low_pct:grad_dec]
+    grad_center = np.full((random.randint(1, 6)), grad_low_pct)
+    grad_grid = np.hstack((grad_grid, grad_center, np.flip(grad_grid)))
+    grad_grid = np.tile(grad_grid,(height,1))
+    grad_list.append(grad_grid)
+
+    # Create Standard Bar with Low Dark Wide Light
+    grad_dec = (grad_low_pct-grad_high_pct)/(line_width)
+    grad_grid = np.mgrid[grad_high_pct:grad_low_pct:grad_dec]
+    grad_exterior = np.full((random.randint(1, 6)), grad_high_pct)
+    grad_grid = np.hstack((grad_grid, np.flip(grad_grid), grad_exterior))
+    grad_grid = np.tile(grad_grid,(height,1))
+    grad_list.append(grad_grid)
+
+    mask = random.choice(grad_list)
+    while (mask.shape[1] < width):
+      mask = np.hstack((mask, random.choice(grad_list)))
+
+    return mask[:,0:width]
+
