+++ src/Augraphy/Augmentations/noisetexturize.py
@@ -1,0 +1,88 @@
+################################################################################
+# File: noisetexture.py
+#
+# Class: NoiseTexturizeAugmentation
+#
+# Description: This file contains a class defining an Augmentation which creates
+#               a random noise based texture pattern to emulate paper textures.
+################################################################################
+
+
+
+################################################################################
+# Imports
+################################################################################
+
+import cv2
+import numpy as np
+import random
+
+
+
+################################################################################
+# Class Definition
+################################################################################
+
+class NoiseTexturizeAugmentation(Augmentation):
+  """
+      Consequently applies noise patterns to the original image from big to small.
+
+      sigma: defines bounds of noise fluctuations
+      turbulence: defines how quickly big patterns will be replaced with the small ones. The lower
+      value - the more iterations will be performed during texture generation.
+  """
+
+  def __init__(self, sigma_range=(3, 10), turbulence_range=(2, 5), probability=0.5, layer='paper'):
+    super().__init__(probability=probability)
+    self.sigma_range = sigma_range
+    self.turbulence_range = turbulence_range
+    self.layer = layer
+
+  # Constructs a string representation of this Augmentation.
+  def __repr__(self):
+    return f"NoiseTexturizeAugmentation(sigma_range={self.sigma_range}, turbulence_range={self.turbulence_range}, probability={self.probability})"
+
+  # Applies the Augmentation to input data.
+  def __call__(self, data, force=False):
+    if (force or self.should_run()):
+      image = data[self.layer][-1].result
+
+      sigma = random.randint(self.sigma_range[0], self.sigma_range[1])
+      turbulence = random.randint(self.turbulence_range[0], self.turbulence_range[1])
+
+      result = image.astype(float)
+      cols, rows, ch = image.shape
+      ratio = cols
+      while not ratio == 1:
+          result += self.noise(cols, rows, ratio, sigma=sigma)
+          ratio = (ratio // turbulence) or 1
+      cut = np.clip(result, 0, 255)
+
+      data[self.layer].append(AugmentationResult(self, cut.astype(np.uint8)))
+
+  def noise(self, width, height, ratio, sigma):
+      """
+      The function generates an image, filled with gaussian nose. If ratio parameter is specified,
+      noise will be generated for a lesser image and then it will be upscaled to the original size.
+      In that case noise will generate larger square patterns. To avoid multiple lines, the upscale
+      uses interpolation.
+
+      :param ratio: the size of generated noise "pixels"
+      :param sigma: defines bounds of noise fluctuations
+      """
+      mean = 0
+      #assert width % ratio == 0, "Can't scale image with of size {} and ratio {}".format(width, ratio)
+      #assert height % ratio == 0, "Can't scale image with of size {} and ratio {}".format(height, ratio)
+
+      h = int(height / ratio)
+      w = int(width / ratio)
+
+      if (h == 0):
+        h = 1
+      if (w == 0):
+        w = 1
+
+      result = np.random.normal(mean, sigma, (w, h, 1))
+      if ratio > 1:
+        result = cv2.resize(result, dsize=(width, height), interpolation=cv2.INTER_LINEAR)
+      return result.reshape((width, height, 1))
